+++ changelog.md
@@ -1,12 +1,25 @@
 # CHANGELOG
 
-## Unreleased
+## unreleased
+
+### Refactor
+
+* [3332](https://github.com/zeta-chain/node/pull/3332) - implement orchestrator V2. Move BTC observer-signer to V2
+* [3360](https://github.com/zeta-chain/node/pull/3360) - update protocol contract imports using consolidated path
+* [3349](https://github.com/zeta-chain/node/pull/3349) - implement new bitcoin rpc in zetaclient with improved performance and observability
 
 ### Fixes
 
+* [3374](https://github.com/zeta-chain/node/pull/3374) - remove minimum rent exempt check for SPL token withdrawals
 * [3348](https://github.com/zeta-chain/node/pull/3348) - add support to perform withdraws in ZetaChain `onRevert` call
 
 ## v25.0.0
+
+## Unreleased
+
+## Refactor
+
+* [2863](https://github.com/zeta-chain/node/pull/2863) - refactor zetacore to delete matured ballots and add a migration script to remove all old ballots.
 
 ### Features
 
@@ -27,7 +40,6 @@
 * [3170](https://github.com/zeta-chain/node/pull/3170) - revamp TSS package in zetaclient
 * [3291](https://github.com/zeta-chain/node/pull/3291) - revamp zetaclient initialization (+ graceful shutdown)
 * [3319](https://github.com/zeta-chain/node/pull/3319) - implement scheduler for zetaclient
-* [3332](https://github.com/zeta-chain/node/pull/3332) - implement orchestrator V2. Move BTC observer-signer to V2
 
 ### Fixes